requirements.txt
flask
mysql.connector
requests
locustio